Backend

+++ Backend/controllers/imageController.js
@@ -13,34 +13,51 @@ function newLock() {
   return [lock, unlock];
 }
 
-const getAllImages =  async (req, res) => {
-  const images = await Images.find({});
+const getAllImages = async (req, res) => {
+  const images = await Images.find();
   // console.log(images);
 
-  var [lock, unlock] = newLock();
+  if (images.length === 0) {
+    res.status(404).json({ message: "No files found!" });
+  }
 
-  var imageLinks = [];
+  else {
 
-  let count = 0;
+    var [lock, unlock] = newLock();
 
-  for (let image of images) {
+    var imageLinks = [];
+
+    let count = 0;
+
+    for (let image of images) {
       imageUrl = await getObjectSignedUrl(image.image)
       imageLinks.push({ imageURL: imageUrl })
       count++;
 
       if (count === images.length)
-          unlock();
+        unlock();
+    }
+
+    await lock;
+
+    res.status(200).send(imageLinks)
   }
-
-  await lock;
-
-  res.status(200).send(imageLinks)
 }
 
 const getSingleImage = async (req, res) => {
+  console.log(req.params.id);
+  const image = await Images.find({ image: req.params.id });
+  if (image.length === 0) {
+    res.status(404).json({ message: "File not found!" });
+  }
 
-  imageUrl = await getObjectSignedUrl(req.params.id);
-  res.status(200).json({ imageURL: imageUrl })
+  else {
+
+    console.log(image);
+    imageUrl = await getObjectSignedUrl(req.params.id);
+
+    res.status(200).json({ imageURL: imageUrl })
+  }
 }
 
 const uploadImage = async (req, res) => {
@@ -50,8 +67,8 @@ const uploadImage = async (req, res) => {
   const imageName = generateFileName()
 
   const fileBuffer = await sharp(file.buffer)
-      .resize({ height: 1920, width: 1080, fit: "contain" })
-      .toBuffer()
+    .resize({ height: 1920, width: 1080, fit: "contain" })
+    .toBuffer()
 
   await uploadFile(fileBuffer, imageName, file.mimetype)
 
@@ -62,12 +79,25 @@ const uploadImage = async (req, res) => {
 
 const deleteImage = async (req, res) => {
   const id = req.params.id
-  const image = await Images.find({image:id})
+  const image = await Images.find({ image: id })
 
-  await deleteFile(id)
+  console.log(image)
 
-  await Images.deleteOne({id:image.id})
-  res.send(image)
+  if (image.length === 0) {
+    console.log("hi")
+    res.status(404).json({ message: "File not found!" });
+    // throw new Error("File Not Found!")
+  }
+
+  else {
+
+    console.log("hello")
+
+    await deleteFile(id)
+
+    await Images.deleteOne({ image: id })
+    res.status(200).send(image)
+  }
 }
 
-module.exports = {getAllImages,getSingleImage,uploadImage,deleteImage};
+module.exports = { getAllImages, getSingleImage, uploadImage, deleteImage };
